lab10/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.3.0
beanie==1.29.0
blinker==1.9.0
Brotli==1.1.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
ConfigArgParse==1.7
cryptography==44.0.3
dnspython==2.7.0
ecdsa==0.19.1
exceptiongroup==1.2.2
fastapi==0.115.11
Flask==3.1.1
flask-cors==5.0.1
Flask-Login==0.6.3
gevent==25.5.1
geventhttpclient==2.3.3
greenlet==3.2.2
h11==0.14.0
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
lazy-model==0.2.0
locust==2.34.0
MarkupSafe==3.0.2
motor==3.7.0
msgpack==1.1.0
passlib==1.7.4
psutil==7.0.0
pyasn1==0.4.8
pycparser==2.22
pydantic==2.10.6
pydantic-mongo==3.1.0
pydantic_core==2.27.2
pymongo==4.12.0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
pyzmq==26.4.0
requests==2.32.3
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.46.1
toml==0.10.2
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.4.0
uvicorn==0.34.0
Werkzeug==3.1.3
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
